python/requirements.txt
pycuda
nvidia-tensorrt
numpy
